requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
dnspython==2.1.0
mongoengine==0.23.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.8.0
pymongo==3.12.1
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
